requirements.txt
streamlit==1.28.0
pandas==2.1.0
plotly==5.17.0
scikit-learn==1.3.0
numpy==1.25.0
openpyxl==3.1.2

# PostgreSQL ve Veritabanı Bağlantısı
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0